+++ packages/zosfiles/src/methods/create/Create.ts
@@ -234,20 +234,10 @@ export class Create {
                             tempOptions.blksize = tempOptions.lrecl;
                         }
 
-                        if(tempOptions.blksize  <= tempOptions.lrecl ){
+                        if(tempOptions.blksize <= tempOptions.lrecl ){
                             tempOptions.blksize = tempOptions.lrecl;
-                            if(tempOptions.recfm === null || tempOptions.recfm === undefined){
-                                tempOptions.recfm = "FB";
-                            }
-                            switch (tempOptions.recfm.toUpperCase()) {
-                                case "V":
-                                case "VB":
-                                case "VBS":
-                                case "VS":
-                                    tempOptions.blksize += 4;
-                                    break;
-                                default:
-                                    break;
+                            if (tempOptions.recfm && tempOptions.recfm.toUpperCase().startsWith("V")) {
+                                tempOptions.blksize += 4;
                             }
                         }
                         break;
@@ -315,31 +305,8 @@ export class Create {
                         break;
 
                     case "recfm":
-                    // zOSMF defaults to F if missing so mimic it's behavior
-                        if (tempOptions.recfm === null || tempOptions.recfm === undefined) {
-                            tempOptions.recfm = "F";
-                        }
+                    // no validation
 
-                        // F, V, or U are required; B, A, M, S, T or additional
-                        // VBA works on mainframe but not via zOSMF
-                        switch (tempOptions.recfm.toUpperCase()) {
-                            case "D":
-                            case "DB":
-                            case "DBS":
-                            case "DS":
-                            case "F":
-                            case "FB":
-                            case "FBS":
-                            case "FS":
-                            case "V":
-                            case "VB":
-                            case "VBS":
-                            case "VS":
-                            case "U":
-                                break;
-                            default:
-                                throw new ImperativeError({ msg: ZosFilesMessages.invalidRecfmOption.message + tempOptions.recfm });
-                        }
                         break;
 
                     // SMS class values